Remove AnnotatedServiceFace and move AnnotatedService to public.
