+++ core/src/main/java/com/linecorp/armeria/internal/server/annotation/HttpResultUtil.java
@@ -23,6 +23,7 @@ import com.linecorp.armeria.common.MediaType;
 import com.linecorp.armeria.common.ResponseHeaders;
 import com.linecorp.armeria.common.ResponseHeadersBuilder;
 import com.linecorp.armeria.server.ServiceRequestContext;
+import com.linecorp.armeria.server.annotation.AnnotatedService;
 import com.linecorp.armeria.server.annotation.HttpResult;
 
 final class HttpResultUtil {
