+++ core/src/main/java/com/linecorp/armeria/server/annotation/AnnotatedService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 LINE Corporation
+ * Copyright 2017 LINE Corporation
  *
  * LINE Corporation licenses this file to you under the Apache License,
  * version 2.0 (the "License"); you may not use this file except in compliance
@@ -13,23 +13,23 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package com.linecorp.armeria.server;
+
+package com.linecorp.armeria.server.annotation;
 
 import java.lang.reflect.Method;
 
 import com.linecorp.armeria.common.HttpStatus;
-import com.linecorp.armeria.internal.server.annotation.AnnotatedService;
-import com.linecorp.armeria.server.annotation.ExceptionHandler;
-import com.linecorp.armeria.server.annotation.Path;
-import com.linecorp.armeria.server.annotation.StatusCode;
+import com.linecorp.armeria.server.HttpService;
+import com.linecorp.armeria.server.Route;
+import com.linecorp.armeria.server.ServerBuilder;
 
 /**
  * An {@link HttpService} which is defined by a {@link Path} or HTTP method annotations.
- * {@link AnnotatedServiceFacade} is intended for public access.
- * Please check out the documentation at
+ * This class is designed to provide a common interface for {@link AnnotatedService}
+ * internally. Please check out the documentation at
  * <a href="https://armeria.dev/docs/server-annotated-service">Annotated HTTP Service</a> to use this.
  */
-public interface AnnotatedServiceFacade extends HttpService {
+public interface AnnotatedService extends HttpService {
 
     /**
      * Returns the annotated service object specified with {@link ServerBuilder#annotatedService(Object)}.
@@ -55,7 +55,24 @@ public interface AnnotatedServiceFacade extends HttpService {
     HttpStatus defaultStatus();
 
     /**
-     * Returns an {@link HttpService} decorated with an {@link ExceptionHandler}.
+     * Returns service name for this {@link AnnotatedService}.
      */
-    HttpService withExceptionHandler(HttpService service);
+    String serviceName();
+
+    /**
+     * Returns a boolean value indicating whether the
+     * serviceName is specified by the {@link ServiceName}.
+     */
+    boolean serviceNameSetByAnnotation();
+
+    /**
+     * Returns method name which is annotated in {@link AnnotatedService}.
+     */
+    String methodName();
+
+    /**
+     * Returns a unique ID to distinguish overloaded methods.
+     * If the method is not overloaded, it should return 0.
+     */
+    int overloadId();
 }
